requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
sqlparse==0.4.4
tzdata==2024.1
